+++ examples/physics_events.rs
@@ -61,7 +61,7 @@ fn handle_physics_events(trigger: Trigger<TiledColliderCreated>) {
 #[derive(Component)]
 struct MyCustomPhysicsComponent;
 
-#[derive(Default, Clone)]
+#[derive(Default, Clone, Reflect)]
 struct MyCustomPhysicsBackend;
 
 // This simple example will just spawn an entity with a `MyCustomPhysicsComponent` Component,