+++ src/physics/avian.rs
@@ -24,7 +24,7 @@ use crate::prelude::*;
 /// App::new()
 ///     .add_plugins(TiledPhysicsPlugin::<TiledPhysicsAvianBackend>::default());
 /// ```
-#[derive(Default, Clone)]
+#[derive(Default, Clone, Reflect)]
 pub struct TiledPhysicsAvianBackend;
 
 impl TiledPhysicsBackend for TiledPhysicsAvianBackend {